requirements.txt
ib-insync>=0.9.86
pandas>=2.0.0
pyarrow>=16.0.0
python-dotenv>=1.0.1



